requirements.txt
Django==1.9.6
Pillow==2.3.0
django-angular==0.8.2
django-bower==5.1.0
django-celery==3.1.17
django-filter==0.13.0
django-zurb-foundation==5.5.0
djangorestframework==3.3.3
flake8==2.5.4
geographiclib==1.46.3
geopy==1.11.0
polycircles==0.1
redis==2.10.3
requests==2.2.1
simplekml==1.3.0
simpy==3.0.9